# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Creator Engine API",
    "description": "Central multi-tenant Telegram platform. All amounts are in minor units.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/owner/resources/{resource}": {
      "get": {
        "summary": "Resources",
        "operationId": "resources_api_owner_resources__resource__get",
        "parameters": [
          {
            "name": "resource",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Resource"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "after",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "After"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/owner/tenants/{tenant_id}/action": {
      "post": {
        "summary": "Tenant Action",
        "operationId": "tenant_action_api_owner_tenants__tenant_id__action_post",
        "parameters": [
          {
            "name": "tenant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tenant Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OwnerAction"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/owner/saas-plans/{plan_id}": {
      "put": {
        "summary": "Saas Plan",
        "operationId": "saas_plan_api_owner_saas_plans__plan_id__put",
        "parameters": [
          {
            "name": "plan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Plan Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaaSPlanUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/owner/tenants/{tenant_id}/features/{key}": {
      "put": {
        "summary": "Feature Flag",
        "operationId": "feature_flag_api_owner_tenants__tenant_id__features__key__put",
        "parameters": [
          {
            "name": "tenant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tenant Id"
            }
          },
          {
            "name": "key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FeatureUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "FeatureUpdate": {
        "properties": {
          "enabled": {
            "type": "boolean",
            "title": "Enabled"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "enabled"
        ],
        "title": "FeatureUpdate"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "OwnerAction": {
        "properties": {
          "action": {
            "type": "string",
            "enum": [
              "SUSPEND",
              "REACTIVATE"
            ],
            "title": "Action"
          },
          "reason": {
            "type": "string",
            "maxLength": 500,
            "minLength": 5,
            "title": "Reason"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "action",
          "reason"
        ],
        "title": "OwnerAction"
      },
...
      "SaaSPlanUpdate": {
        "properties": {
          "amount_xtr": {
            "type": "integer",
            "maximum": 10000.0,
            "minimum": 1.0,
            "title": "Amount Xtr"
          },
          "bots": {
            "type": "integer",
            "maximum": 1000.0,
            "minimum": 1.0,
            "title": "Bots"
          },
          "admins": {
            "type": "integer",
            "maximum": 1000.0,
            "minimum": 1.0,
            "title": "Admins"
          },
          "active_contacts": {
            "type": "integer",
            "maximum": 10000000.0,
            "minimum": 1.0,
            "title": "Active Contacts"
          },
          "campaigns_month": {
            "type": "integer",
            "maximum": 10000.0,
            "minimum": 0.0,
            "title": "Campaigns Month"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "amount_xtr",
          "bots",
          "admins",
          "active_contacts",
          "campaigns_month"
        ],
        "title": "SaaSPlanUpdate"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}